refactor: move to react wrapper for slides
Note: the webpack config is currently tied to a commit hash; we'll want to fix that once this (https://github.com/objectpartners/object-partners-presentation-config/pull/6) is merged

feat: get code highlighting working; fix extra space in code blocks

chore: minor refactor
