--- slidedeck/src/app/app.component.pug
@@ -1,1 +1,0 @@
-opi-slide-deck([slides]="slides")
