--- slidedeck/src/app/shared/components/button/button.component.html
@@ -1,1 +1,0 @@
-<ng-content select="button"></ng-content>
