--- slidedeck/src/app/shared/components/display-date/display-date.component.scss
@@ -1,7 +1,0 @@
-.buttons {
-  display: flex;
-  margin: 1rem 0;
-  opi-button {
-    margin: 0 8px;
-  }
-}
