--- slidedeck/src/app/shared/components/dog-list/dog-list.component.html
@@ -1,6 +1,0 @@
-<ul class="dog-list">
-  <li class="dog" *ngFor="let dog of dogs">
-    <h3>{{dog.name}}</h3>
-    <p>{{dog.breed}} ({{dog.age}})</p>
-  </li>
-</ul>
