--- slidedeck/src/app/shared/components/dog-list/dog-list.component.scss
@@ -1,9 +1,0 @@
-.dog-list {
-  list-style-type: none;
-  padding: 0;
-  display: block;
-  margin: 0;
-  .dog {
-    text-align: left;
-  }
-}
