--- slidedeck/src/app/shared/components/index.ts
@@ -1,5 +1,0 @@
-export { ButtonComponent } from './button/button.component';
-export { DisplayDateComponent } from './display-date/display-date.component';
-export { DogListComponent } from './dog-list/dog-list.component';
-export { ObservableComponent } from './observable/observable.component';
-export { UpperLowerComponent } from './upper-lower/upper-lower.component';
