--- slidedeck/src/app/shared/components/observable/observable.component.html
@@ -1,9 +1,0 @@
-<div id="observable-area" #observableArea>
-  <div class="container" [ngSwitch]="!!(position.x && position.y)">
-    <span class="position" *ngSwitchCase="true">
-      <span>X: {{position.x}}</span>
-      <span>Y: {{position.y}}</span>
-    </span>
-    <span *ngSwitchDefault>Mouse over me</span>
-  </div>
-</div>
