--- slidedeck/src/app/shared/components/observable/observable.component.scss
@@ -1,19 +1,0 @@
-$observable-bg-color: mix(white, red, 25%);
-
-#observable-area {
-  height: 50vh;
-  width: 50vh;
-  background: linear-gradient(to bottom, $observable-bg-color, red);
-  margin: 0 auto;
-  position: relative;
-  z-index: 2;
-  .container {
-    margin: 0;
-    color: white;
-    text-align: center;
-    line-height: 50vh;
-  }
-  .position {
-    pointer-events: none;
-  }
-}
