--- slidedeck/src/app/shared/components/observable/observable.interface.ts
@@ -1,4 +1,0 @@
-export interface IMousePosition {
-  x?: number;
-  y?: number;
-}
