--- slidedeck/src/app/shared/components/upper-lower/upper-lower.component.scss
@@ -1,7 +1,0 @@
-.buttons {
-  display: flex;
-  opi-button {
-    margin: 0 8px;
-  }
-}
-
