--- slidedeck/src/app/shared/components/upper-lower/upper-lower.component.ts
@@ -1,18 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'opi-upper-lower',
-  templateUrl: './upper-lower.component.html',
-  styleUrls: ['./upper-lower.component.scss']
-})
-export class UpperLowerComponent {
-  strValue: string = 'Hello World';
-
-  pipe: string;
-
-  constructor() {}
-
-  setPipe(pipe: string) {
-    this.pipe = pipe;
-  }
-}
