--- slidedeck/src/app/slide-deck/slide-deck.component.pug
@@ -1,11 +1,0 @@
-.reveal
-  .slides
-    section(data-state="title")
-      h1 React
-      h2 Nebraska.Code() - May 2017
-    section(*ngFor="let slideDeck of slides;")
-      section(*ngFor="let slide of slideDeck")
-        dynamic-html([content]="slide")
-    section(data-state="title")
-      h1 Thanks!
-      h2 Any questions?
