+++ slidedeck/src/components/slide-deck/slide-deck.tsx
@@ -1,0 +1,95 @@
+import * as React from 'react';
+import * as renderHtml from 'react-render-html';
+
+import 'reveal.js/css/reveal.css';
+import 'object-partners-revealjs-theme';
+import 'highlight.js/styles/monokai.css';
+
+export interface SlideDeckProps {
+  slides: any[][];
+}
+
+export interface SlideDeckState {}
+
+export class SlideDeck extends React.Component<SlideDeckProps, SlideDeckState> {
+  componentDidMount() {
+    require.ensure([
+      'reveal.js',
+      'reveal.js/lib/js/classList.js',
+      'reveal.js/lib/js/head.min.js',
+      'reveal.js/lib/js/html5shiv.js',
+      'highlight.js'
+    ], () => {
+      const Reveal = require('reveal.js');
+      require('reveal.js/lib/js/classList.js');
+      require('reveal.js/lib/js/head.min.js');
+      require('reveal.js/lib/js/html5shiv.js');
+
+      (window as any).Reveal = Reveal;
+      const hljs = require('highlight.js');
+
+      Reveal.initialize({
+        history: true,
+        margin: 0.20,
+        dependencies: [
+          {
+            async: true,
+            src: require('reveal.js/plugin/zoom-js/zoom.js')
+          },
+          {
+            async: true,
+            src: require('reveal.js/plugin/markdown/marked.js')
+          },
+          {
+            async: true,
+            src: require('reveal.js/plugin/notes/notes.js')
+          },
+          {
+            src: '',
+            callback() {
+              hljs.initHighlightingOnLoad();
+            }
+          }
+        ]
+      });
+
+    });
+  }
+
+  render() {
+    const { slides } = this.props;
+    return (
+      <div className="reveal">
+        <div className="slides">
+          <section data-state="title">
+            <h1>React</h1>
+            <h2>Nebraska.Code() - May 2017</h2>
+          </section>
+          {
+            slides
+              .map((deck: any[], deckIndex: number) => {
+                return (
+                  <section key={deckIndex}>
+                    {
+                      deck
+                        .map((slide: any, slideIndex: number) => {
+                          return (
+                            <section key={slideIndex}>
+                              {renderHtml(slide)}
+                            </section>
+                          );
+                        })
+                    }
+                  </section>
+                );
+              })
+          }
+          <section data-state="title">
+            <h1>Thanks!</h1>
+            <h2>Any questions?</h2>
+          </section>
+        </div>
+      </div>
+    );
+  }
+}
