--- slidedeck/src/main.ts
@@ -1,10 +1,0 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { AppModule } from './app/app.module';
-
-if ( process.env.NODE_ENV === 'production' ) {
-  enableProdMode();
-}
-
-platformBrowserDynamic().bootstrapModule(AppModule);
