+++ slidedeck/src/public/index.pug
@@ -4,5 +4,13 @@ html
     meta(charset="utf-8")
     meta(name="viewport", content="width=device-width, initial-scale=1")
     title React Redux Workshop | Object Partners, Inc.
+    style(type="text/css").
+      #slide-root {
+        position: absolute;
+        top: 0;
+        right: 0;
+        bottom: 0;
+        left: 0;
+      }
   body
-    app-root
+    div#slide-root
