+++ slidedeck/src/slides/05-react-forms/00-react-forms.pug
@@ -83,9 +83,7 @@ section
         employee: this.state.employee,
         errors: this.state.errors
       });
-    }
-
-    </code>
+    }</code>
 
 section
   h1 lab prep
