+++ slidedeck/tsconfig.json
@@ -4,6 +4,7 @@
     "module": "commonjs",
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
+    "jsx": "react",
     "sourceMap": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
